+++ Test/Integration/OAuthTest.php
@@ -19,7 +19,7 @@ class OAuthTest extends Setup
             'clientId' => 'client_id$development$integration_client_id'
         ));
         $gateway->oauth()->createTokenFromCode(array(
-            'code' => 'integration_oauth_auth_code_' . rand(0, 299)
+            'code' => 'integration_oauth_auth_code_'.rand(0, 299)
         ));
     }
 
@@ -280,5 +280,20 @@ class OAuthTest extends Setup
         parse_str($queryString, $query);
 
         $this->assertEquals(array('credit_card', 'paypal'), $query['payment_methods']);
+    }
+
+
+    public function testComputeSignature()
+    {
+        $gateway = new Braintree\Gateway(array(
+            'clientId' => 'client_id$development$integration_client_id',
+            'clientSecret' => 'client_secret$development$integration_client_secret'
+        ));
+
+        $urlToSign = 'http://localhost:3000/oauth/connect?business%5Bname%5D=We+Like+Spaces&client_id=client_id%24development%24integration_client_id';
+
+        $signature = $gateway->oauth()->computeSignature($urlToSign);
+
+        $this->assertEquals('a36bcf10dd982e2e47e0d6a2cb930aea47ade73f954b7d59c58dae6167894d41', $signature);
     }
 }