+++ src/Braintree/Configuration.php
@@ -235,16 +235,7 @@ class Configuration
      */
     public function baseUrl()
     {
-        static $defaultPorts = array(
-            'http' => 80,
-            'https' => 443,
-        );
-
-        if ($this->portNumber() === $defaultPorts[$this->protocol()]) {
-            return sprintf('%s://%s', $this->protocol(), $this->serverName());
-        } else {
-            return sprintf('%s://%s:%d', $this->protocol(), $this->serverName(), $this->portNumber());
-        }
+        return sprintf('%s://%s:%d', $this->protocol(), $this->serverName(), $this->portNumber());
     }
 
     /**