+++ src/Braintree/Transaction/EuropeBankAccountDetails.php
@@ -1,0 +1,25 @@
+<?php
+namespace Braintree\Transaction;
+
+use Braintree\Instance;
+
+/**
+ * Europe bank account details from a transaction
+ * Creates an instance of europe bank account details as returned from a transaction
+ *
+ * @package    Braintree
+ * @subpackage Transaction
+ * @copyright  2014 Braintree, a division of PayPal, Inc.
+ *
+ * @property-read string $accountHolderName
+ * @property-read string $bic
+ * @property-read string $imageUrl
+ * @property-read string $mandateAcceptedAt
+ * @property-read string $mandateReferenceNumber
+ * @property-read string $maskedIban
+ * @property-read string $token
+ * @uses Braintree_Instance inherits methods
+ */
+class EuropeBankAccountDetails extends Instance
+{
+}